+++ axum-live-view/Cargo.toml
@@ -18,6 +18,7 @@ axum = { version = "0.4", features = ["ws"] }
 axum-live-view-macros = { path = "../axum-live-view-macros", version = "0.1" }
 bytes = "1.0"
 futures-util = { version = "0.3", default-features = false, features = ["alloc"] }
+http = "0.2"
 percent-encoding = "2.1"
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
