+++ axum-live-view/src/extract.rs
@@ -1,0 +1,141 @@
+use crate::{
+    html::Html,
+    life_cycle::{process_view_messages, EmbedLiveView, ViewTaskHandle},
+};
+use async_trait::async_trait;
+use axum::{
+    extract::{
+        ws::{self, WebSocket, WebSocketUpgrade},
+        FromRequest, RequestParts,
+    },
+    http::{HeaderMap, Uri},
+    response::{IntoResponse, Response},
+};
+use futures_util::{
+    sink::SinkExt,
+    stream::{StreamExt, TryStreamExt},
+};
+use serde::{de::DeserializeOwned, Serialize};
+use std::fmt::Debug;
+
+use self::rejection::LiveViewUpgradeRejection;
+
+#[derive(Debug)]
+pub struct LiveViewUpgrade {
+    inner: LiveViewUpgradeInner,
+}
+
+#[derive(Debug)]
+enum LiveViewUpgradeInner {
+    Http,
+    Ws(Box<(WebSocketUpgrade, Uri, HeaderMap)>),
+}
+
+#[async_trait]
+impl<B> FromRequest<B> for LiveViewUpgrade
+where
+    B: Send,
+{
+    type Rejection = LiveViewUpgradeRejection;
+
+    async fn from_request(req: &mut RequestParts<B>) -> Result<Self, Self::Rejection> {
+        if let Ok(ws) = WebSocketUpgrade::from_request(req).await {
+            let uri = req.uri().clone();
+            let headers = req.headers().cloned().ok_or_else(|| {
+                LiveViewUpgradeRejection::HeadersAlreadyExtracted(Default::default())
+            })?;
+
+            Ok(Self {
+                inner: LiveViewUpgradeInner::Ws(Box::new((ws, uri, headers))),
+            })
+        } else {
+            Ok(Self {
+                inner: LiveViewUpgradeInner::Http,
+            })
+        }
+    }
+}
+
+impl LiveViewUpgrade {
+    pub fn response<F, M>(self, gather_view: F) -> Response
+    where
+        F: FnOnce(EmbedLiveView<'_, M>) -> Html<M>,
+        M: Serialize + DeserializeOwned + Send + 'static,
+    {
+        let (ws, uri, headers) = match self.inner {
+            LiveViewUpgradeInner::Http => {
+                let embed = EmbedLiveView::noop();
+                return gather_view(embed).into_response();
+            }
+            LiveViewUpgradeInner::Ws(data) => *data,
+        };
+
+        let mut view = None;
+
+        let embed = EmbedLiveView::new(&mut view);
+
+        gather_view(embed);
+
+        let view = if let Some(view) = view {
+            view
+        } else {
+            return ws.on_upgrade(|_| async {}).into_response();
+        };
+
+        ws.on_upgrade(|socket| run_view_on_socket(socket, view, uri, headers))
+            .into_response()
+    }
+}
+
+async fn run_view_on_socket<M>(
+    socket: WebSocket,
+    view: ViewTaskHandle<M>,
+    uri: Uri,
+    headers: HeaderMap,
+) where
+    M: Serialize + DeserializeOwned + Send + 'static,
+{
+    let (write, read) = socket.split();
+
+    let write = write.with(|msg| async move {
+        let encoded_msg = ws::Message::Text(serde_json::to_string(&msg)?);
+        Ok::<_, anyhow::Error>(encoded_msg)
+    });
+    futures_util::pin_mut!(write);
+
+    let read = read
+        .map_err(anyhow::Error::from)
+        .and_then(|msg| async move {
+            if let ws::Message::Text(text) = msg {
+                serde_json::from_str(&text).map_err(Into::into)
+            } else {
+                anyhow::bail!("received message from socket that wasn't text")
+            }
+        });
+    futures_util::pin_mut!(read);
+
+    if let Err(err) = process_view_messages(write, read, view, uri, headers).await {
+        tracing::error!(%err, "encountered while processing socket");
+    }
+}
+
+pub mod rejection {
+    use axum::{
+        extract::rejection::HeadersAlreadyExtracted,
+        response::{IntoResponse, Response},
+    };
+
+    #[derive(Debug)]
+    #[non_exhaustive]
+    pub enum LiveViewUpgradeRejection {
+        HeadersAlreadyExtracted(HeadersAlreadyExtracted),
+    }
+
+    impl IntoResponse for LiveViewUpgradeRejection {
+        fn into_response(self) -> Response {
+            match self {
+                Self::HeadersAlreadyExtracted(inner) => inner.into_response(),
+            }
+        }
+    }
+}
