+++ examples/chat/src/main.rs
@@ -7,7 +7,11 @@ use axum::{
     AddExtensionLayer, Router,
 };
 use axum_live_view::{
-    html, js_command, live_view, EventData, Html, LiveView, LiveViewUpgrade, SelfHandle, Updated,
+    event_data::EventData,
+    html, js_command,
+    life_cycle::SelfHandle,
+    live_view::{self, Updated},
+    Html, LiveView, LiveViewUpgrade,
 };
 use serde::{Deserialize, Serialize};
 use std::{
@@ -26,18 +30,6 @@ async fn main() {
     tracing_subscriber::fmt::init();
 
     let messages: Messages = Default::default();
-
-    // {
-    //     let pubsub = pubsub.clone();
-    //     let messages = messages.clone();
-    //     let mut new_messages = pubsub.subscribe(&NewMessageTopic).await.unwrap();
-    //     tokio::spawn(async move {
-    //         while let Some(Json(msg)) = new_messages.next().await {
-    //             messages.lock().unwrap().push(msg);
-    //             let _ = pubsub.broadcast(&ReRenderMessageList, ()).await;
-    //         }
-    //     });
-    // }
 
     let (tx, _) = broadcast::channel::<NewMessagePing>(1024);
